face-service/requirements.txt
face_recognition==1.3.0
flask==3.0.3
Pillow>=10.4.0
numpy>=1.26.4
